requirements.txt
absl-py==1.1.0
Adafruit-Blinka==6.15.0
adafruit-circuitpython-busdevice==5.1.2
adafruit-circuitpython-motor==3.3.5
adafruit-circuitpython-pca9685==3.4.1
adafruit-circuitpython-register==1.9.8
adafruit-circuitpython-servokit==1.3.6
Adafruit-PlatformDetect==3.19.6
Adafruit-PureIO==1.1.9
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
contextvars==2.4
cycler==0.11.0
Cython==0.29.28
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
docker-pycreds==0.4.0
entrypoints==0.4
evdev==1.4.0
flake8==4.0.1
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.18
google-auth==2.8.0
google-auth-oauthlib==0.4.6
graphsurgeon==0.4.5
grpcio==1.46.3
idna==3.3
idna-ssl==1.1.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jetcam==0.0.0
jetracer==0.0.0
Jetson.GPIO==2.0.17
Jinja2==3.0.3
json5==0.9.8
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
kaleido==0.2.1
kiwisolver==1.3.1
line-profiler==3.4.0
Mako==1.1.6
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
mpu6050-raspberrypi==1.2
mpu9250-jmdev==1.0.12
multidict==5.2.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
picotui==1.2.1
Pillow==9.0.1
pkg_resources==0.0.0
platformdirs==2.4.0
plotly==5.8.2
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycuda==2020.1
pyDeprecate==0.3.1
pyflakes==2.4.0
pyftdi==0.53.3
Pygments==2.11.2
pynput==1.7.6
pyparsing==3.0.7
pyrsistent==0.18.0
pyserial==3.5
python-dateutil==2.8.2
python-xlib==0.31
pytools==2022.1
pytorch-lightning==1.5.10
pytz==2022.1
pyusb==1.2.1
PyYAML==6.0
pyzmq==23.0.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
Send2Trash==1.8.0
sentry-sdk==1.5.6
shortuuid==1.0.8
simple-pid==1.0.1
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
tenacity==8.0.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorrt==8.0.1.6
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
timm==0.5.4
torch==1.10.0
torch2trt==0.3.0
torchmetrics==0.8.2
torchvision==0.11.1
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
typing_extensions==4.1.1
uff==0.6.9
urllib3==1.26.8
wandb==0.12.10
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
yarl==1.7.2
yaspin==2.1.0
zipp==3.6.0
